Refactors and improves trajectory analysis pipeline
Improves the trajectory analysis pipeline by refactoring code for better readability and maintainability.
Adds new filtering steps based on position and box range to refine the selection of key trajectory points.
Implements logic to classify and group non-alternating sequences of start and end indices and selects representatives based on acceleration, resulting in a cleaner and more meaningful dataset.
Includes basic checks and raises errors when number of reach segments are not 16.
Adds utilities to plot x, position and speed with segments for reach validation